+++ .changeset/ninety-bags-dream.md
@@ -2,4 +2,4 @@
 '@nordeck/matrix-poll-widget': patch
 ---
 
-Uniformity the diagrams colors of the answer options for the poll result chart in light and dark theme
+Unify the diagrams colors of the answer options for the poll result chart in light and dark theme.
