+++ src/components/PollResultModal/PollResultChart/PollResultChartBar.test.tsx
@@ -23,7 +23,7 @@ import { ComponentType, PropsWithChildren } from 'react';
 import { mockPoll } from '../../../lib/testUtils';
 import { PollType } from '../../../model';
 import { PollInvalidAnswer, SelectPollResults } from '../../../store';
-import { PollResultChartBar } from './PollResultChartBar';
+import PollResultChartBar from './PollResultChartBar';
 
 const mockSimpleBarChart = jest.fn();
 
@@ -215,98 +215,6 @@ describe('<PollResultChartBar/>', () => {
         color: {
           scale: {
             No: '#CBB701',
-            Yes: '#8AB3FF',
-          },
-        },
-        axes: {
-          bottom: {
-            mapsTo: 'group',
-            scaleType: 'labels',
-          },
-          left: {
-            mapsTo: 'value',
-            ticks: {
-              values: [0, 1, 2],
-            },
-          },
-        },
-        height: '270px',
-        theme: 'g100',
-        toolbar: {
-          enabled: false,
-        },
-        tooltip: {
-          enabled: false,
-        },
-      },
-    });
-  });
-
-  it('should repeat the colors if the answers are more then 7', () => {
-    jest
-      .mocked(useThemeSelection)
-      .mockReturnValue({ theme: 'dark', setTheme: jest.fn(), isModal: false });
-    const pollResult: SelectPollResults = {
-      poll: mockPoll({
-        content: {
-          pollType: PollType.Open,
-          startTime: '2020-01-01T03:33:55Z',
-          answers: [
-            { id: '1', label: 'Yes' },
-            { id: '2', label: 'No' },
-            { id: '3', label: 'Maybe' },
-            { id: '4', label: 'Of course' },
-            { id: '5', label: "I don't know yet" },
-            { id: '6', label: 'Not sure' },
-            { id: '7', label: 'Probably Yes' },
-            { id: '8', label: 'Probably No' },
-          ],
-        },
-      }),
-      results: {
-        votes: {
-          '@user-1': '1',
-          '@user-2': PollInvalidAnswer,
-          '@user-3': '2',
-          '@user-4': '1',
-          '@user-5': '5',
-          '@user-6': PollInvalidAnswer,
-          '@user-7': '5',
-          '@user-8': '4',
-        },
-      },
-      votingRights: ['@user-1', '@user-2', '@user-3', '@user-4'],
-    };
-
-    render(<PollResultChartBar isFinished pollResults={pollResult} />, {
-      wrapper: Wrapper,
-    });
-
-    expect(screen.getByText('SimpleBarChart')).toBeInTheDocument();
-
-    expect(mockSimpleBarChart).toBeCalledWith({
-      data: [
-        { group: 'Yes', value: 2 },
-        { group: 'No', value: 1 },
-        { group: 'Maybe', value: 0 },
-        { group: 'Of course', value: 1 },
-        { group: "I don't know yet", value: 2 },
-        { group: 'Not sure', value: 0 },
-        { group: 'Probably Yes', value: 0 },
-        { group: 'Probably No', value: 0 },
-        { group: 'Invalid', value: 2 },
-      ],
-      options: {
-        color: {
-          scale: {
-            "I don't know yet": '#40BFBD',
-            Invalid: '#CBB701',
-            Maybe: '#C29EFF',
-            No: '#CBB701',
-            'Not sure': '#EB8995',
-            'Of course': '#F684BB',
-            'Probably No': '#8AB3FF',
-            'Probably Yes': '#ED905E',
             Yes: '#8AB3FF',
           },
         },
